requirements.txt
aiosqlite==0.17.0
annotated-types==0.6.0
anyio==3.7.1
bcrypt==4.0.1
certifi==2023.11.17
click==8.1.7
dnspython==2.4.2
email-validator==2.1.0.post1
fastapi==0.104.1
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
idna==3.6
iso8601==1.1.0
Pillow==10.1.0
pydantic==2.5.2
pydantic_core==2.14.5
PyJWT==2.8.0
pypika-tortoise==0.1.6
python-multipart==0.0.6
pytz==2023.3.post1
sniffio==1.3.0
starlette==0.27.0
tortoise-orm==0.20.0
typing_extensions==4.8.0
uvicorn==0.24.0.post1
fiona==1.9.5
numpy==1.24.3
opencv-python==4.8.1.78
opencv-contrib-python==4.8.1.78
Pillow==10.0.1
pyproj==3.6.1
rasterio==1.3.9
reportlab==4.0.7
shapely==2.0.2
ultralytics==8.0.221